requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.3
pydantic>=2.5.0
